add the logic of model id deletion while deleting the key
